セルクラスをモノクロセルとして定義して別のファイルに移した
